Removed the concept of color in the registration messages
The color is not used by server anymore so it is not necessary. Also
removed all the XMessage classes. They could be replaced with a few
simple lines of code in the client class. In this case I think less is more and it makes
things easier to understand. The XMessage class also added some fields
that were not necessary, e.g., the receiving player id.