Refactored SnakeClient to improve test-stability and reduce layers.
Basically removed the intermediate web socket layer and moved that code
into SnakeClient. Rewrote the tests to use the StubWebSocket and a
SyncrhonousTaskScheduler to get rid of the Retry stuff. This makes the
tests pass. They were very unstable before.

Merged all SnakeTests into a single functional test to reduce the
complexity.

Now, the SnakeClient requires a socket on construction. The caller is
responsible for connecting the socket before starting any game.