Removed event classes and cleaned up Map
The event classes contained a lot of properties not used anywhere. I
thought it would be nice to only parse the data that we actually use.
Therefore I could not see the value for the event classes anymore. The
Client handles the printing on incoming messages, the bot is only
responsible for generating a new move.

The map is now parsed in the FromJson methods. The reason for this is that
it made it easier to parse the incoming data into MapCoordinate instances
instead of array positions. Therefore, I could also convert some of the
methods to properties. E.g., GetFoods => FoodPostions. This made some of
the test cases obsolete.